Main update to UI and file structure to look neater
1. Added header to all front-end files
2. Updated index page to look better
3. updated login page to look better
4. Added left padding to make all pages display better.
5. removed clear.jsp, close.jsp, release.jsp, run.jsp, unrelease.jsp, but kept their functionalities within the display.jsp page. Works fine.
6. Changed the table in display.jsp, made the buttons to have a separate table from the poll itself.

ISSUE:
see message in discord. (关于display.jsp 和 vote.jsp显示结果时的选项错位)

Comment:
After testing, found a possible bug: when the choice contains space at the end, it might not record the vote correctly.

Things in the requirement but not done: optional description.

Things not in the requirement but seems better if we do it: 让用户可以自定义有几个选项。但是看起来修改这个需要修改很多现有的逻辑，做起来工作量比较大，老师也没说必须要，所以只是提一下，不一定做。